+++ src/app/modules/main/pages/profile/profile.component.ts
@@ -1,9 +1,10 @@
-import { Router } from '@angular/router';
-import { AuthService } from './../../../shared/services/auth.service';
-import { UsersService } from './../../../shared/services/users.service';
 import { Component, OnInit } from '@angular/core';
 import { FormControl, FormGroup, Validators } from '@angular/forms';
+import { Router } from '@angular/router';
 import Swal from 'sweetalert2';
+
+import { AuthService } from './../../../shared/services/auth.service';
+import { UsersService } from './../../../shared/services/users.service';
 
 @Component({
   selector: 'app-profile',
@@ -16,6 +17,7 @@ export class ProfileComponent implements OnInit {
     Validators.required,
     Validators.email,
   ]);
+
   formUser = new FormGroup({
     id: new FormControl(this.user$.id, Validators.required),
     name: new FormControl(this.user$.name, Validators.required),
@@ -23,7 +25,7 @@ export class ProfileComponent implements OnInit {
       Validators.required,
       Validators.email,
     ]),
-    password: new FormControl('', Validators.required),
+    password: new FormControl(null, Validators.required),
   });
 
   constructor(
@@ -33,15 +35,15 @@ export class ProfileComponent implements OnInit {
   ) {}
 
   ngOnInit(): void {
-    console.log(this.user$);
   }
 
   saveUser() {
     const data = this.formUser.value;
-    const id = data.id
+    const id = data.id;
+
     this.usersService.put(id, data).subscribe((resp) => {
       this.user$ = resp;
-      this.authService.storeUser(this.user$)
+      this.authService.storeUser(this.user$);
 
       Swal.fire({
         position: 'top-end',
@@ -53,18 +55,29 @@ export class ProfileComponent implements OnInit {
     });
   }
 
-  deleteUser(id: string){
-    this.usersService.delete(id).subscribe( (resp) => {
-      this.authService.removeSession();
-      this.router.navigate(['login']);
+  deleteUser(id: string) {
+    Swal.fire({
+      title: 'Tem certeza que deseja excluir sua conta?',
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#9C27B0',
+      cancelButtonColor: '#F44336',
+      confirmButtonText: 'Sim, excluir minha conta!',
+      cancelButtonText: 'Cancelar',
+    }).then((result) => {
+      if (result.isConfirmed) {
+        this.usersService.delete(id).subscribe((resp) => {
+          this.authService.removeSession();
+          this.router.navigate(['login']);
+        });
 
-      Swal.fire({
-        position: 'top-end',
-        icon: 'success',
-        title: `Conta excluída com sucesso!`,
-        showConfirmButton: false,
-        timer: 1500,
-      });
-    })
+        Swal.fire({
+          position: 'top-end',
+          title: 'Conta excluída!',
+          icon: 'success',
+          showConfirmButton: false,
+        });
+      }
+    });
   }
 }
